Move JAVA_BIN to systemd service files
The EnvironmentFile paths are config locations so if a user ever
modified them, they're not replaced. By defining this in the service
file, the packaging will update the location. The result is that even
if a java location changes that a simple yum/apt upgrade will respect
the new value in most situations.